fixes naming after moved to groups
