+++ src/groups/RouteNetworkDiagram/RouteNetworkDiagram.scss
@@ -1,4 +1,4 @@
-.identify-feature-page {
+.route-network-diagram {
   width: 100%;
 
   .feature-information-container {
