+++ src/groups/RouteNetworkDiagram/RouteNetworkDiagramGql.ts
