+++ src/pages/IdentifyFeaturePage/index.ts
@@ -1,3 +1,4 @@
 import IdentifyFeaturePage from "./IdentifyFeaturePage";
+import "./IdentifyFeaturePage.scss";
 
 export default IdentifyFeaturePage;
